requirements.txt
# Core dependencies
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pillow==10.0.0

# Enhanced document analysis
PyMuPDF==1.22.5  # aka fitz - advanced PDF processing with better image extraction

# LLM API dependencies
# Using the latest versions for better compatibility
openai>=1.3.0
httpx>=0.24.0
anthropic==0.5.0
google-generativeai==0.3.1

# Optional OCR capability
# pytesseract==0.3.10  # Uncomment if OCR needed (requires Tesseract installation)